+++ frontend/voice-assistant/src/components/ProductCard/ProductCard.jsx
@@ -45,7 +45,7 @@ export default function ProductCard(product) {
         {/* <span className='btn-layout'> */}
             <BtnOutlined name_of_btn={"В избранное"} icon_for_btn={heartIconWh} />
             <BtnFilled name_of_btn={"Купить"} icon_for_btn={carttIconWh} />
-            {/* </span> */}
+        {/* </span> */}
         </div>
     )
 }
